+++ src/Resources/assets/js/contao-flatpickr-bundle-theme.js
@@ -1,2 +1,3 @@
 import '../../public/assets/flatpickr.css';
+import 'flatpickr/dist/plugins/monthSelect/style.css';
 import '../scss/flatpickr-bundle.scss';